engine: rename the nodejs engine service to javascript
