+++ cmd/app/start/start.go
@@ -274,7 +274,7 @@ func (s *Start) CreateStartCommand() *cobra.Command {
 			BoolP(
 				"disable-docker", "D",
 				s.configs.DisableDocker,
-				"Used to run horusec without docker if enabled it will only run the following tools: horusec-csharp, horusec-kotlin, horusec-java, horusec-kubernetes, horusec-leaks, horusec-nodejs, horusec-dart, horusec-nginx. Example: -D=\"true\"",
+				"Used to run horusec without docker if enabled it will only run the following tools: horusec-csharp, horusec-kotlin, horusec-java, horusec-kubernetes, horusec-leaks, horusec-javascript, horusec-dart, horusec-nginx. Example: -D=\"true\"",
 			)
 	}
