+++ internal/services/engines/javascript/rule_manager.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package nodejs
+package javascript
 
 import (
 	engine "github.com/ZupIT/horusec-engine"
